dll_39: (int) cast for u8 cmp flips cmplwi to cmpwi (99.71→99.97% fuzzy)

+++ src/main/dll/FRONT/dll_39.c
@@ -410,7 +410,7 @@ void n_attractmode_prepareMovie(void)
         VIWaitForRetrace();
         lbl_803DD64D = NATTRACTMODE_MOVIE_RETRACE_COUNTDOWN;
         lbl_803DD698 = 0;
-        if (lbl_803DD614 == NATTRACTMODE_MOVIE_STATE_RELEASED) {
+        if ((int)lbl_803DD614 == NATTRACTMODE_MOVIE_STATE_RELEASED) {
           fn_80117B68(NATTRACTMODE_MOVIE_START_FRAME_ALTERNATE,1);
         }
         else {
